refactor: organize files, import paths and remove unused files
