+++ src/config/mod.rs
@@ -1,0 +1,7 @@
+pub mod keystore;
+pub mod loader;
+pub mod types;
+pub mod validation;
+
+// Re-export main types for easier access
+pub use types::*;
