+++ src/key_management/mod.rs
@@ -1,0 +1,8 @@
+pub mod add;
+pub mod delete;
+pub mod list;
+
+// Re-export main functions for easier access
+pub use add::add_key;
+pub use delete::delete_key;
+pub use list::list_keys;
