+++ src/middleware/mod.rs
@@ -1,0 +1,3 @@
+pub mod security;
+
+pub use security::security_headers_middleware;
