+++ src/middleware/security.rs
@@ -1,0 +1,27 @@
+/// Middleware to add security headers to all responses
+pub async fn security_headers_middleware(
+    request: axum::http::Request<axum::body::Body>,
+    next: axum::middleware::Next,
+) -> axum::response::Response {
+    let mut response = next.run(request).await;
+
+    // Add security headers
+    let headers = response.headers_mut();
+    headers.insert("X-Content-Type-Options", "nosniff".parse().unwrap());
+    headers.insert("X-Frame-Options", "DENY".parse().unwrap());
+    headers.insert("X-XSS-Protection", "1; mode=block".parse().unwrap());
+    headers.insert(
+        "Strict-Transport-Security",
+        "max-age=31536000; includeSubDomains".parse().unwrap(),
+    );
+    headers.insert(
+        "Content-Security-Policy",
+        "default-src 'none'".parse().unwrap(),
+    );
+    headers.insert("Referrer-Policy", "no-referrer".parse().unwrap());
+
+    // Remove server information
+    headers.remove("server");
+
+    response
+}
