+++ src/services/mod.rs
@@ -1,0 +1,3 @@
+pub mod signing;
+
+pub use signing::SigningService;
