+++ src/signer/mod.rs
@@ -1,0 +1,9 @@
+pub mod starknet_signer;
+pub mod transaction_hash;
+
+#[cfg(test)]
+mod tests;
+
+// Re-export main types for easier access
+pub use starknet_signer::StarknetSigner;
+pub use transaction_hash::compute_transaction_hash;
